What I attempted
This is the app I tried to make over the summer. I ceased work on this
project because it didn’t make any business sense to make a pink
keyboard when companies like SwiftKey already have their own upcoming
version of a pink keyboard for iOS 8. I couldn’t compete with an
established brand.

But I did learn a lot about making a mobile application in the process
so that counts for something.